# OOCL.Image.Client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "OOCL.Image API",
    "description": "API + WebApp using OpenCL Kernels for image generation etc.",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "github: alarmclock-kisser",
      "email": "marcel.king91299@gmail.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/OpenCl/status": {
      "get": {
        "tags": [
          "OpenCl"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OpenClServiceInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenClServiceInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenClServiceInfo"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/OpenCl/devices": {
      "get": {
        "tags": [
          "OpenCl"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OpenClDeviceInfo"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OpenClDeviceInfo"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OpenClDeviceInfo"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/OpenCl/initialize-id": {
      "post": {
        "tags": [
          "OpenCl"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 2
              }
            },
            "text/json": {
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 2
              }
            },
            "application/*+json": {
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 2
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OpenClServiceInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenClServiceInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenClServiceInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/OpenCl/initialize-name": {
      "post": {
        "tags": [
          "OpenCl"
        ],
        "parameters": [
          {
            "name": "deviceName",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "Core"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OpenClServiceInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenClServiceInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenClServiceInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/OpenCl/release": {
      "delete": {
        "tags": [
          "OpenCl"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OpenClServiceInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenClServiceInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenClServiceInfo"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/OpenCl/kernel-infos": {
      "get": {
        "tags": [
          "OpenCl"
        ],
        "parameters": [
          {
            "name": "onlyCompiled",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OpenClKernelInfo"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OpenClKernelInfo"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OpenClKernelInfo"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/OpenCl/execute-on-image": {
      "post": {
        "tags": [
          "OpenCl"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteOnImageRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteOnImageRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteOnImageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/OpenCl/execute-create-image": {
      "post": {
        "tags": [
          "OpenCl"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateImageRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateImageRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateImageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageObjDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ImageObjDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "info": {
            "$ref": "#/components/schemas/ImageObjInfo"
          },
          "data": {
            "$ref": "#/components/schemas/ImageObjData"
          }
        },
        "additionalProperties": false
      },
...
      "OpenClDeviceInfo": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "integer",
            "format": "int32"
          },
          "deviceName": {
            "type": "string",
            "nullable": true
          },
          "deviceType": {
            "type": "string",
            "nullable": true
          },
          "platformName": {
            "type": "string",
            "nullable": true
          },
          "localMemoryBytes": {
            "type": "string",
            "nullable": true
          },
          "globalMemoryBytes": {
            "type": "string",
            "nullable": true
          },
          "addressBits": {
            "type": "string",
            "nullable": true
          },
          "compilerAvailable": {
            "type": "string",
            "nullable": true
          },
          "driverVersion": {
            "type": "string",
            "nullable": true
          },
          "maxClockFrequency": {
            "type": "string",
            "nullable": true
          },
          "maxComputeUnits": {
            "type": "string",
            "nullable": true
          },
          "maxWorkGroupSize": {
            "type": "string",
            "nullable": true
          },
          "maxWorkItemDimensions": {
            "type": "string",
            "nullable": true
          },
          "maxWorkItemSizes": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "string",
            "nullable": true
          },
          "imageSupport": {
            "type": "string",
            "nullable": true
          },
          "image2DMaxSize": {
            "type": "string",
            "nullable": true
          },
          "image3DMaxSize": {
            "type": "string",
            "nullable": true
          },
          "vendor": {
            "type": "string",
            "nullable": true
          },
          "vendorId": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OpenClKernelInfo": {
        "type": "object",
        "properties": {
          "argumentsCount": {
            "type": "integer",
            "format": "int32"
          },
          "argumentNames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "argumentType": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "inputPointerName": {
            "type": "string",
            "nullable": true
          },
          "outputPointerName": {
            "type": "string",
            "nullable": true
          },
          "inputPointerType": {
            "type": "string",
            "nullable": true
          },
          "outputPointerType": {
            "type": "string",
            "nullable": true
          },
          "mediaType": {
            "type": "string",
            "nullable": true
          },
          "functionName": {
            "type": "string",
            "nullable": true
          },
          "colorInputArgNames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "pointersCount": {
            "type": "integer",
            "format": "int32"
          },
          "needsImage": {
            "type": "boolean"
          },
          "compiledSuccessfully": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OpenClServiceInfo": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "integer",
            "format": "int32"
          },
          "deviceName": {
            "type": "string",
            "nullable": true
          },
          "platformName": {
            "type": "string",
            "nullable": true
          },
          "initialized": {
            "type": "boolean"
          },
          "currentDevice": {
            "$ref": "#/components/schemas/OpenClDeviceInfo"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}